requirements.txt
Flask==1.1.2
Flask-Session==0.3.1
psycopg2-binary
SQLAlchemy==1.3.16
python-dotenv==0.12.0
WTForms==2.2.1
passlib==1.7.2
Werkzeug==0.16.0